move /pages/common/Button to /components
